Move testCreateExternalTableWithInaccessibleSchemaLocation to TestExternalHiveTable
